requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
pyarrow>=12.0.0
pydantic>=2.0.0
pytest>=7.4.0
pyyaml>=6.0.0
